+++ packages/modules/src/modules/epics/ContributionsEpic.stories.tsx
@@ -660,7 +660,8 @@ WithNewsletterSignup.args = {
         highlightedText: undefined,
         heading: 'Sign up to First Edition, our free daily newsletter',
         newsletterSignup: {
-            url: 'https://www.theguardian.com/email/form/plaintone/morning-briefing',
+            url:
+                'https://www.theguardian.com/email/form/plaintone/morning-briefing?utm_campaign=epic',
         },
     },
 };
